Completed essential refactor of server actions into services, fixed grad plan issues
