+++ app/auth/authorize/page.tsx
@@ -1,79 +1,40 @@
-'use client';
+import { Suspense } from 'react';
+import { redirect } from 'next/navigation';
+import { cookies } from 'next/headers';
+import { createServerClient } from '@supabase/ssr';
+import { AuthorizeClient } from '@/components/auth/authorize-client';
 
-import { useState, Suspense } from 'react';
-import { useRouter, useSearchParams } from 'next/navigation';
-import { createSupabaseBrowserClient } from '@/lib/supabase/client';
-import AuthorizationPopup from '@/components/auth/authorization-popup';
+export const dynamic = 'force-dynamic'; // ensure we always evaluate auth on request
 
-// Separate component that uses useSearchParams and is wrapped in <Suspense>
-function AuthorizeContent() {
-  const [showPopup] = useState(true);
-  const [isLoading, setIsLoading] = useState(false);
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const supabase = createSupabaseBrowserClient();
-
-  const handleAgree = async () => {
-    setIsLoading(true);
-    try {
-      const { data: { user } } = await supabase.auth.getUser();
-      if (!user) {
-        router.push('/login');
-        return;
+async function getUser() {
+  const cookieStore = await cookies();
+  const supabase = createServerClient(
+    process.env.NEXT_PUBLIC_SUPABASE_URL!,
+    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
+    {
+      cookies: {
+        getAll() { return cookieStore.getAll(); },
+        setAll(cookiesToSet) {
+          cookiesToSet.forEach(({ name, value, options }) => {
+            cookieStore.set(name, value, options);
+          });
+        }
       }
-      const { error } = await supabase
-        .from('profiles')
-        .update({
-          authorization_agreed: true,
-          authorization_agreed_at: new Date().toISOString()
-        })
-        .eq('id', user.id);
-      if (error) {
-        console.error('Error updating authorization:', error);
-        alert('Error saving authorization. Please try again.');
-        setIsLoading(false);
-        return;
-      }
-      const next = searchParams.get('next') || '/dashboard';
-      router.push(next);
-    } catch (error) {
-      console.error('Error in handleAgree:', error);
-      alert('An error occurred. Please try again.');
-      setIsLoading(false);
     }
-  };
-
-  const handleGoBack = () => {
-    supabase.auth.signOut();
-    router.push('/login');
-  };
-
-  return (
-    <div className="min-h-screen bg-gray-50">
-      <AuthorizationPopup
-        isOpen={showPopup}
-        onAgree={handleAgree}
-        onGoBack={handleGoBack}
-        schoolName="BYU"
-      />
-      {isLoading && (
-        <div className="fixed inset-0 bg-black/50 flex items-center justify-center z-[60]">
-          <div className="bg-white rounded-lg p-6">
-            <div className="flex items-center gap-3">
-              <div className="animate-spin rounded-full h-6 w-6 border-b-2 border-primary" />
-              <span className="text-lg">Processing...</span>
-            </div>
-          </div>
-        </div>
-      )}
-    </div>
   );
+  const { data: { user } } = await supabase.auth.getUser();
+  return user;
 }
 
-export default function AuthorizePage() {
+export default async function AuthorizePage() {
+  const user = await getUser();
+  if (!user) {
+    redirect('/login?next=/auth/authorize');
+  }
+
   return (
     <Suspense fallback={<div className="min-h-screen flex items-center justify-center">Loading...</div>}>
-      <AuthorizeContent />
+      <AuthorizeClient schoolName="BYU" defaultNext="/dashboard" />
     </Suspense>
   );
 }
